requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0  # For DeepSeek API compatibility
requests>=2.31.0
python-dateutil>=2.8.2
aiohttp>=3.9.1
asyncio>=3.4.3
duckduckgo-search>=4.1.1 